Add falling rain effect
Added the initial rain effect. Rain starts at a random position on the
top of the screen, falls and accelerates, and is reset when it hits the
bottom of the screen.

Switched the int fields out for floats in the Point and Raindrop types
to allow for smooth motion.

+++ src/Point.hh
@@ -5,9 +5,9 @@
  */
 class Point {
 public:
-	int x;
-	int y;
+	float x;
+	float y;
 
 public:
-	Point(int x = 0, int y = 0): x(x), y(y) {}
+	Point(float x = 0.0F, float y = 0.0F): x(x), y(y) {}
 };
